requirements.txt
Flask==3.1.0
flask_sqlalchemy==3.1.1
Pillow==11.2.1
pytesseract==0.3.13
